Moved client code to couchbase cloud directory/package
